+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.1.2 (2025-03-23)
+
+- feat: Added Arch Linux AUR package support (`yay -S s-screen`)
+- docs: Added website documentation link to package
+
 ## v0.1.2 (2025-03-21)
 
 - chore: Automatic releases